requirements.uv.txt
altair==5.5.0
anyio==4.8.0
anywidget==0.9.13
asttokens==3.0.0
attrs==25.1.0
bokeh==3.6.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
docutils==0.21.2
drawdata==0.3.7
executing==2.2.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
h11==0.14.0
idna==3.10
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab-widgets==3.0.13
kiwisolver==1.4.8
llvmlite==0.44.0
marimo==0.10.19
markdown==3.7
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.24.2
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
polars==1.21.0
prompt-toolkit==3.0.50
psutil==6.1.1
psygnal==0.11.1
pure-eval==0.2.3
pyarrow==19.0.0
pycrdt==0.11.1
pygments==2.19.1
pymdown-extensions==10.14.3
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.22.3
ruff==0.9.4
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
widgetsnbextension==4.0.13
xyzservices==2025.1.0
